refactor: use MailboxHttpAttachmentService in PreviewService
- use an attachment service that is able to route request to correct mailbox
